Fix main Swift file budget failure
