+++ Sources/AgentResumeScriptStore.swift
@@ -1,0 +1,69 @@
+import Foundation
+
+enum AgentResumeScriptStore {
+    private static let directoryName = "cmux-agent-resume"
+    private static let scriptTTL: TimeInterval = 24 * 60 * 60
+
+    static func writeLauncherScript(
+        command: String,
+        kind: RestorableAgentKind,
+        sessionId: String,
+        fileManager: FileManager,
+        temporaryDirectory: URL,
+        returnToLoginShell: Bool = false,
+        workingDirectory: String? = nil
+    ) -> URL? {
+        let directoryURL = temporaryDirectory.appendingPathComponent(directoryName, isDirectory: true)
+        do {
+            try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true)
+            try? fileManager.setAttributes([.posixPermissions: 0o700], ofItemAtPath: directoryURL.path)
+            pruneOldScripts(in: directoryURL, fileManager: fileManager)
+
+            let safeSessionPrefix = sessionId
+                .prefix(12)
+                .map { character -> Character in
+                    character.isLetter || character.isNumber || character == "-" ? character : "_"
+                }
+            let scriptURL = directoryURL.appendingPathComponent(
+                "\(kind.rawValue)-\(String(safeSessionPrefix))-\(UUID().uuidString).zsh",
+                isDirectory: false
+            )
+            var lines = [
+                "#!/bin/zsh",
+                "rm -f -- \"$0\" 2>/dev/null || true"
+            ]
+            if returnToLoginShell {
+                lines.append(contentsOf: TerminalStartupReturnShellScript.commandThenReturnLines(
+                    command: command,
+                    workingDirectory: workingDirectory
+                ))
+            } else {
+                lines.append(command)
+            }
+            let contents = lines.joined(separator: "\n") + "\n"
+            try contents.write(to: scriptURL, atomically: true, encoding: .utf8)
+            try? fileManager.setAttributes([.posixPermissions: 0o600], ofItemAtPath: scriptURL.path)
+            return scriptURL
+        } catch {
+            return nil
+        }
+    }
+
+    private static func pruneOldScripts(in directoryURL: URL, fileManager: FileManager) {
+        guard let scriptURLs = try? fileManager.contentsOfDirectory(
+            at: directoryURL,
+            includingPropertiesForKeys: [.contentModificationDateKey],
+            options: [.skipsHiddenFiles]
+        ) else {
+            return
+        }
+
+        let cutoff = Date().addingTimeInterval(-scriptTTL)
+        for scriptURL in scriptURLs where scriptURL.pathExtension == "zsh" {
+            let values = try? scriptURL.resourceValues(forKeys: [.contentModificationDateKey])
+            if let modified = values?.contentModificationDate, modified < cutoff {
+                try? fileManager.removeItem(at: scriptURL)
+            }
+        }
+    }
+}
